+++ PostHog/Replay/UIView+Util.swift
@@ -1,0 +1,28 @@
+//
+//  UIView+Util.swift
+//  PostHog
+//
+//  Created by Manoel Aranda Neto on 21.03.24.
+//
+
+#if os(iOS)
+    import Foundation
+    import UIKit
+
+    extension UIView {
+        func isVisible() -> Bool {
+            if isHidden || alpha == 0 || frame == .zero {
+                return false
+            }
+            return true
+        }
+
+        func isNoCapture() -> Bool {
+            if let identifier = accessibilityIdentifier {
+                return identifier.lowercased().contains("ph-no-capture")
+            }
+
+            return false
+        }
+    }
+#endif