+++ PostHogExample/AppDelegate.swift
@@ -18,12 +18,13 @@ class AppDelegate: NSObject, UIApplicationDelegate {
         config.captureScreenViews = false
         config.captureApplicationLifecycleEvents = false
         config.flushAt = 1
-        config.flushIntervalSeconds = 10
+        config.flushIntervalSeconds = 30
         config.debug = true
+        config.sendFeatureFlagEvent = false
+        config.sessionReplay = true
 
         PostHogSDK.shared.setup(config)
-//        PostHogSDK.shared.debug()
-//        PostHogSDK.shared.capture("App started!")
+        PostHogSDK.shared.debug()
 
         let defaultCenter = NotificationCenter.default
 
@@ -38,6 +39,6 @@ class AppDelegate: NSObject, UIApplicationDelegate {
     }
 
     @objc func receiveFeatureFlags() {
-        print("receiveFeatureFlags")
+        print("user receiveFeatureFlags callback")
     }
 }