+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- iOS session recording - very first alpha release [#115](https://github.com/PostHog/posthog-ios/pull/115)
+
 ## 3.2.4 - 2024-03-12
 
 - `maxQueueSize` wasn't respected when capturing events [#116](https://github.com/PostHog/posthog-ios/pull/116)